+++ doc/release_notes.rst
@@ -10,6 +10,10 @@ Release Notes
 Upcoming Release
 ================
 
+* Made central heating supply temperatures dynamic based on an adaptation of a reference curve from Pieper et al. (2019) (https://www.sciencedirect.com/science/article/pii/S0360544219305857?via%3Dihub).
+
+* Added option to use country-specific district heating forward and return temperatures. Defaults to lower temperatures in Scandinavia.
+
 * Added unsustainable biomass potentials for solid, gaseous, and liquid biomass. The potentials can be phased-out and/or
   substituted by the phase-in of sustainable biomass types using the config parameters
   ``biomass: share_unsustainable_use_retained`` and ``biomass: share_sustainable_potential_available``.